More faithfully copied acb's original algorithm. randomTitleTwo() closely emulates the probability in acb's algorithm that a given term will appear. Additionally efforts were made to improve the randomness of selected terms, however work must continue to create more dynamic results.
As part of this effort, added randomPlusPlus() function as a replacement to the Math.random() call in a futile effort to increase randomness, making use of the randomBytes package to improve randomization. However some terms are still appearing repetitively.

Additioanlly created a new function twoAdjectives() to generate two random elements from the adjective array and check that they are dissimilar. This acts as a replacement to two direct calls to randomArrayIndex() with the adjective array as an argument which often produced repetitive results. The check that the adjectives are distinct could be improved.

trimE function now only trims the inputted string of 'e'. 'ist' or 'ism' is added to the new term in the calling function as in acb's original design. Added a special case to trimE to account for inputted strings ending in 'ist' to avoid output such as 'individualistism' despite the interesting neologisms this implies.

Accounted for 'Foucaldian' adjective as opposed to 'Foucaltist' in termAboutIntellectual()

Added a random ideology to the abstNounsTwo array which increases the likelihood of the occurrence of any given ideology.