Moved randomization of terms array into main.newTerm().
Added additional tracking to account for combination ideologies such as 'Marxism-Leninism-etc.ism' in scraper.js which were previously counted as prefixes. Added randomIsm() to return a random marxism-leninism to main.newTerm()

In both scraper.getIdeologies and scraper.getArtMovements, moved the loop that checks if a movement ends with another movement and the loop that checks for unhyphenated prefixes to a while loop which ends if no more prefixes or movements are to be derived. Implemented this improvement after discovering an additional iteration of the 'movement ends with other movement' loop uncovered more derivatives of 'culturalism' in the ideologies array.

Added tracking of hypens and more intelligent use of hyphenization based on the frequency of hyphens for each given prefix in the source dataset.

Expanded depth of Cheerio search to capture all ideologies including 'Hoppean anarcho-capitalism' though the query is now quite broad and any significant revisions to the Wikipedia page could cause errors in runtime.

Increased likelihood that any ideology or artMovement appears by adding them to array of terms and adding helper function trimIsm()

randomArrayIndex now shuffles randomly every array it encounters.