Modified scraper.js to account for capitalization of terms. When a term is first encountered its case is preserved as-is. But if a lowercase version of the term is encountered, this version is preferred, with the assumption that terms which are proper nouns (such as 'Marxism') will never be lower-case.
This required the replacement of the native array.includes() function by the custom function main.containsCaseInsensitive() which acts as a case-insensitive array.includes().

The function pushLowerCase() was also added which pushes a term to the array only if the new term is an identically spelled and lower-case version of an existing term.

Modified main.titleCase() so that ideologies processed in main.prefixAndTerm() are not inadvertently lower-cased if not already lower-case.

Modified main.prefixAndTerm() to receive a term as an argument to preserve the case of the inputted term.